% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Generation of simulation data}
\usage{
genData(n0, p, m = 10, case = 1, rho = 0.5, SNR = NULL)
}
\arguments{
\item{n0}{the common sample size for each group}

\item{p}{the number of explanatory variables including intercept}

\item{m}{the number of groups (one of 10, 20 and 50)}

\item{case}{true cluster pattern; the number of the true clusters is
\eqn{m_\ast = 0.3 m} for case 1 and $m_\ast = 0.6 m$ for case 2}

\item{rho}{a parameter of autocorrelation matrix for a matrix of explanatory
variables}

\item{SNR}{SNR value}
}
\value{
a list with the following elements:
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{group}{a vector of group indexes}

\item{adj}{a matrix with two columns expressing adjacency among groups}

\item{Mu}{a vector of true mean}

\item{XI}{a matrix of coefficients for cluster}

\item{BETA}{a matrix of coefficients for group}

\item{cluster}{a list expressing group indexes of each cluster}

\item{cluster.labels}{a vector of cluster indexes}

\item{SNR}{SNR}

\item{SDC}{a dataframe with pairs of groups that they reside different clusters
and standardized difference of coefficients}
}
\description{
\code{genData} This function generates simulation data (v0.1.0)
}
\examples{
#genData(n0, p)
}
