% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JTT.R
\name{JTT}
\alias{JTT}
\title{Consistent clustering for group-wise linear regression}
\usage{
JTT(
  y,
  X,
  group,
  adj,
  PSE = "penalized",
  alpha = NULL,
  pse.alpha = NULL,
  Rcpp = TRUE
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept (intercept is
automatically included)}

\item{group}{a vector of group indexes}

\item{adj}{a matrix with two columns expressing adjacency among groups;
each row is a pair of indexes of adjacent groups}

\item{PSE}{post selection estimation; \code{"OLS"} (default) or \code{"penalized"}}

\item{alpha}{a penalty parameter of \eqn{GC_p} criterion for JTT score;
default (\code{NULL}) is \eqn{HCGC_p} criterion}

\item{pse.alpha}{a penalty parameter of \eqn{GC_p} criterion for post selection
estimation when \code{PSE = "penalized"}; default (\code{NULL}) is \eqn{MC_p} criterion}

\item{Rcpp}{if \code{TRUE}, C++ codes are used}
}
\value{
a list with the following elements:
\item{coefficients}{a matrix of estimated coefficients}

\item{fitted.values}{a vector of fitted values}

\item{cluster}{a list expressing group indexes of each cluster}

\item{cluster.labels}{a vector of cluster indexes}

\item{JTT.scores}{a dataframe with pairs of groups and JTT scores}

\item{PostSelEst}{information of post selection estimation}

\item{summary}{summary of estimation results}
}
\description{
\code{JTT} By join-two-together method, this function performs
clustering for group-wise linear regression with graph structure (v0.1.0)
}
\examples{
#JTT(y, X, group, adj)
}
